- Move FakeCurrentTime to test folder. - Improved tests reducing indirection using FakeCurrentTime. - Scalar tpe declaration
